requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.9.3
aiosignal==1.3.1
anyio==4.6.2.post1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
attrs==24.2.0
awkward==2.6.3
awkward-cpp==32
awkward0==0.15.5
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.9.0
Brotli==1.0.9
cached-property==1.5.2
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cmake==3.29.0.1
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cramjam==2.8.3
cryptography==42.0.2
cycler==0.12.1
debugpy==1.8.9
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
exceptiongroup==1.2.2
executing==2.1.0
fastjsonschema==2.20.0
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.51.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.10.0
gast==0.4.0
gmpy2==2.1.5
google-auth==2.36.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.62.1
h11==0.14.0
h2==4.1.0
h5py==3.10.0
hpack==4.0.0
httpcore==1.0.7
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
iminuit==2.30.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
ipykernel==6.29.5
ipyparallel==9.0.0
ipython==8.29.0
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.28
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
keras==3.2.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
libclang==18.1.1
libprotobuf=3.21.12
lightgbm==4.5.0
lit==18.1.2
llvmlite==0.42.0
lz4==4.3.3
Markdown==3.6
MarkupSafe==3.0.2
matplotlib==3.8.4
matplotlib-inline==0.1.7
metakernel==0.30.2
mistune==3.0.2
ml-dtypes==0.3.2
mpmath==1.3.0
multidict==6.0.5
munkres==1.1.4
namex==0.0.7
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest_asyncio==1.6.0
networkx==3.3
notebook==7.2.2
notebook_shim==0.2.4
numba==0.59.1
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
opt-einsum==3.3.0
optree==0.11.0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.0
parso==0.8.4
patsy==1.0.1
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.3.0
pkgutil_resolve_name==1.3.10
platformdirs==4.3.6
ply==3.11
portalocker==3.0.0
prometheus_client==0.21.0
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==4.25.3
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyaml==24.9.0
pyarrow==11.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
Pygments==2.18.0
PyJWT==2.10.0
pyOpenSSL==24.2.1
pyparsing==3.2.0
PyQt5==5.15.9
PyQt5-sip==12.12.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytz==2024.1
pyu2f==0.1.5
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.21.0
rsa==4.9
scikit-learn==1.4.1.post1
scikit-optimize==0.10.2
scipy==1.14.1
seaborn==0.13.2
Send2Trash==1.8.3
sip==6.7.12
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.2
statsmodels==0.14.4
sympy==1.12
tabulate==0.9.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorflow==2.16.1
tensorflow-estimator=2.11.0
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.4.0
toml==0.10.2
tomli==2.1.0
torch==2.0.1
torch-cluster==1.6.1+pt20cu117
torch-scatter==2.1.1+pt20cu117
torch-sparse==0.6.17+pt20cu117
torch-spline-conv==1.2.2+pt20cu117
torch_geometric==2.5.2
torchaudio==2.0.0
torchvision==0.15.0
tornado==6.4.2
tqdm==4.66.2
traitlets==5.14.3
triton==2.0.0
types-python-dateutil==2.9.0.20241003
typing-utils==0.1.0
typing_extensions==4.12.2
tzdata==2024.2
unicodedata2==15.1.0
uproot==5.3.2
uproot3==3.14.4
uproot3-methods==0.10.1
uri-template==1.3.0
urllib3==2.2.3
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.2
wrapt==1.16.0
xgboost==2.1.2
xrootd==5.5.4
xxhash==3.4.1
yarl==1.9.4
zipp==3.21.0
zstandard==0.23.0